fix: revert changes for expanding lines in arrays
- ## Issue

The alternative proposed by shellcheck that works on both bash and zsh
relies on using the structure

```bash
while IFS='' read -r var; do
  ... something
done < <(command)
```

where `command` is any number of commands to be use in process
substitution.

However the above snippet fails when using fzf due to a bug in zsh
5.8 - 5.9. It is fixed in zsh 5.9 on commit [ddfc81] but it is not
available in ubuntu or other ubunto based distributions.

- ## Resolution

Ways to use a newer version of zsh where attempted but turned out to be
more complexity which may be gone once packages are updated. So it is
decided to rollback the shellcheck suggestion for now and use it again
later on newer releases of ubuntu and friends.

- ## More information

Ref: https://github.com/junegunn/fzf/issues/3947