refactor!(fzf-preview.ps1): Change number of arguments of fzf-preview.ps1
The preview script will change to support only a single argument
which is the path to the item to preview.

Additional information should be delivered through environment variables
from the caller of the script.

This is made to reduce complexity of the script argument parsing as the
bash version has had way less issues by already working under this
concept.