fix(fzf-opts): Set header-border color explicitly
The header-border seems to unsync when launching fzf from (n)vim in
powershell/bash in windows.

If `--color=dark` is used, the header shows all in black colors.
Opmiting it fixes it on the terminal but inside neovim it will show
black. However this behavior is different in linux, so there is not
consitent option to make it all work and setting this in (n)vim when
calling fzf is not a good solution.

This adds the border color as fzf configures it in the dark theme
https://github.com/junegunn/fzf/blob/e15cba0c8c7c9dd3388d260cf5b5de7fc044dfbc/src/tui/tui.go#L899
For some reason that color is lost when using --color=dark inside nvim.